Create uncontrolled canvas type
