+++ apps/frontend/package.json
@@ -25,6 +25,7 @@
         "eslint-plugin-react-hooks": "^4.6.0",
         "js-file-download": "^0.4.12",
         "lodash": "^4.17.20",
+        "lodash.omit": "^4.5.0",
         "mic-recorder-to-mp3": "^2.2.2",
         "prettier": "^3.1.1",
         "react": "^18.2.0",
@@ -86,6 +87,7 @@
         "@rollup/plugin-typescript": "^11.1.5",
         "@types/jest": "^29.5.8",
         "@types/lodash": "^4.14.201",
+        "@types/lodash.omit": "^4.5.9",
         "@types/node": "^20.9.1",
         "@types/react": "^18.2.37",
         "@types/react-datepicker": "^4.19.3",
