+++ apps/frontend/src/components/Fields/SignatureField.tsx
@@ -9,6 +9,7 @@ import { useTranslations } from '../../hooks/useTranslations';
 import { Nullish, Signature, SignaturePoint, Path, PathValue, TranslatedString } from '@3dp4me/types';
 import type ReactSignatureCanvas from 'react-signature-canvas';
 import { NewSiganatureModal, SignatureData } from '../NewSignatureModal/NewSignatureModal';
+import { ControlledSignatureCanvas } from '../ControlledSignatureCanvas/ControlledSignatureCanvas';
 
 export interface SignatureFieldProps<T extends string> {
     displayName: string
@@ -30,59 +31,6 @@ const SignatureField = <T extends string>({
     const [isModalOpen, setIsModalOpen] = useState(false);
     const [isDocumentVisible, setIsDocumentVisible] = useState(false);
     const translations = useTranslations()[0];
-    const sigCanvas = useRef<ReactSignatureCanvas | null>(null);
-
-    useEffect(() => {
-        if (!value?.signatureData || isModalOpen || !sigCanvas.current) return;
-
-        const canvas = sigCanvas.current.getCanvas()
-        if (!canvas) return
-
-        const data = transformSignatureData(
-            canvas,
-            value.signatureData,
-            value.signatureCanvasWidth,
-            value.signatureCanvasHeight,
-        );
-
-        console.log("DATA IS ", data)
-        console.log("CANVAS ID ", sigCanvas.current)
-        sigCanvas.current?.fromData(data as any)
-    }, [value, isModalOpen, sigCanvas.current]);
-
-    /**
-     * Transforms the signature data so that it is able to be displayed on the canvas.
-     * This includes doing some slight manipulation of the structure and scaling
-     */
-    const transformSignatureData = (
-        canvas: HTMLCanvasElement,
-        data: SignaturePoint[][],
-        originalCanvasWidth: number,
-        originalCanvasHeight: number,
-    ) => {
-        const formattedData = data.map((points) => {
-            const withoutColor = points.map((point) => _.omit(point, 'color'));
-            const firstTimestamp = withoutColor.length > 0 ? withoutColor[0].time : 0;
-
-            const padPoints = withoutColor.map((point) => {
-                const scaleFactor = canvas.width / originalCanvasWidth;
-                const deltaT = point.time - firstTimestamp;
-
-                // Scale the data points so that they fit this canvas
-                const x = (point.x / originalCanvasWidth) * canvas.width
-                const y = (point.y / originalCanvasHeight) * canvas.height
-
-                // Scales the time of each touch point.... doens't work great
-                const time = firstTimestamp + deltaT * scaleFactor
-
-                return { x, y, time }
-            });
-
-            return padPoints
-        });
-
-        return formattedData;
-    };
 
     /**
      * Saves signature data points along with the canvas width and height so that
@@ -128,18 +76,6 @@ const SignatureField = <T extends string>({
         );
     };
 
-    /**
-     * Renders the canvas for displaying saved signatures (not for creating new ones).
-     * The library we're using is kind of bad... we can't ever have more than one canvas
-     * elements on screen
-     */
-    const renderCanvas = () => {
-        // We can't render the canvas while the other canvas is visible
-        if (!value?.signatureData || isModalOpen) return null;
-
-        return <SignatureCanvas ref={sigCanvas}/>;
-    };
-
     return (
         <div className="signature-container">
             <h3>{displayName}</h3>
@@ -149,7 +85,7 @@ const SignatureField = <T extends string>({
                 onSave={onNewSignature}
                 isOpen={isModalOpen}/>
             <div className="sig-container">
-                {renderCanvas()}
+                <ControlledSignatureCanvas value={value} />
                 <div className="sig-ctl-container">
                     <Button
                         className="sig-ctl-button doc-btn"
